edited graphql queries in all folder
