+++ client/src/Components/Billing/BillingHistory.js
@@ -1,26 +1,8 @@
 import React from "react";
 import { Query } from 'react-apollo';
-import { gql } from 'apollo-boost';
 import ApolloClient from "apollo-boost";
 import BillingHistoryCard from './BillingHistoryCard';
-
-const GET_BILLING_HISTORY = gql`
-query getBillingHistory($id: ID!){
-    getBillingHistory(id: $id){
-        id
-        date
-        amount_paid
-    }
-  }
-`;
-
-const GET_CURRENT = gql`
-  query getCurrentUser {
-    getCurrentUser {
-      id
-    }
-  }
-`;
+import {GET_CURRENT_USER_QUERY, GET_BILLING_HISTORY} from "../../graphql/queries";
 
 class BillingHistory extends React.Component{
     constructor(props){
@@ -42,7 +24,7 @@ class BillingHistory extends React.Component{
 
         client
         .query({
-            query: GET_CURRENT
+            query: GET_CURRENT_USER_QUERY
         })
         .then(response => {
             this.setState({
