+++ client/src/Components/Billing/BillingPlans.js
@@ -1,5 +1,4 @@
 import React from 'react';
-// import classNames from 'classnames';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 // import AppBar from '@material-ui/core/AppBar';
