+++ client/src/Components/Dashboard/Dashboard.js
@@ -8,7 +8,6 @@ import ExerciseEntry from "./ExerEntry";
 import styled from "styled-components";
 import ApolloClient from "apollo-boost";
 import moment from "moment";
-import gql from "graphql-tag";
 import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 
@@ -23,7 +22,7 @@ import {
 } from "../../graphql/mutations";
 import {
   EXER_QUERY,
-  GET_CURRENT_USERID,
+  GET_CURRENT_USER_QUERY,
   GET_EXERCISE_ENTRIES_QUERY,
   GET_FOOD_ENTRIES_BY_USER_QUERY
 } from "../../graphql/queries";
@@ -73,7 +72,7 @@ const styles = theme => ({
     [theme.breakpoints.down('sm')]: {
       width: "100%",
     },
-    
+
   },
   flexDataConFirst: {
     width: "100%",
@@ -99,36 +98,6 @@ const Hr = styled.div`
   width: 90%;
   height: 1px;
 `;
-
-// const GET_FOOD_ENTRIES_BY_USER_QUERY = gql`
-//   query getFoodEntriesByUserId($userId: ID!) {
-//     getFoodEntriesByUserId(userId: $userId) {
-//       id
-//       date
-//       servingQty
-//       user_id {
-//         username
-//         firstName
-//         lastName
-//         email
-//         id
-//       }
-//       food_id {
-//         id
-//         foodName
-//         caloriesPerServ
-//         fats
-//         proteins
-//         carbs
-//         edamam_id
-//       }
-//       meal_category_id {
-//         id
-//         mealCategoryName
-//       }
-//     }
-//   }
-// `;
 
 class Dashboard extends Component {
   state = {
@@ -157,7 +126,7 @@ class Dashboard extends Component {
     });
     client
       .query({
-        query: GET_CURRENT_USERID
+        query: GET_CURRENT_USER_QUERY
       })
       .then(response => {
         this.setState({
@@ -205,7 +174,7 @@ class Dashboard extends Component {
       });
       client
         .query({
-          query: GET_CURRENT_USERID
+          query: GET_CURRENT_USER_QUERY
         })
         .then(response => {
           this.setState({ currentUser: response.data.getCurrentUser.id });
@@ -306,7 +275,6 @@ class Dashboard extends Component {
           e.target.type === "number" ? parseInt(e.target.value) : e.target.value
       }
     });
-    console.log("exerentry change", this.state.exerEntry);
   };
 
   onFoodEntryChange = e => {
@@ -322,7 +290,6 @@ class Dashboard extends Component {
           e.target.type === "number" ? parseInt(e.target.value) : e.target.value
       }
     });
-    console.log(this.state.foodEntry.foodName);
   };
 
   onFoodChange = e => {
@@ -338,17 +305,6 @@ class Dashboard extends Component {
       }
     });
   };
-  // onMealChange = e => {
-  //   this.setState({
-  //     foodEntry:{
-  //       meal_category_id:{
-  //         ...this.state.foodEntry.meal_category_id,
-  //         [e.target.name]:
-  //           e.target.type === "number" ? parseInt(e.target.value) : e.target.value
-  //       }
-  //     }
-  //   })
-  // }
 
   editFoodEntry = (editId, editEntry, idToken) => {
     const client = new ApolloClient({
@@ -502,7 +458,7 @@ class Dashboard extends Component {
   handleShowFood = () => {
     this.setState({
       showFoodForm: true,
-      selectedFood: {}
+      selectedFood: []
     });
   };
