+++ client/src/Components/Journal/JournalEntry.js
@@ -13,9 +13,8 @@ import { withStyles } from "@material-ui/core/styles";
 import List from "@material-ui/core/List";
 import * as moment from "moment";
 import ApolloClient from "apollo-boost";
-import gql from "graphql-tag";
 
-import { GET_CURRENT_USERID } from "../../graphql/queries";
+import { GET_CURRENT_USER_QUERY, GET_FOOD_BY_ID } from "../../graphql/queries";
 import MealFoods from "./MealFoods";
 
 const styles = theme => ({
@@ -85,14 +84,6 @@ const styles = theme => ({
   }
 });
 
-const GET_FOOD_BY_ID = gql`
-  query getFoodById($foodId: ID!) {
-    getFoodById(foodId: $foodId) {
-      id
-      edamam_id
-    }
-  }
-`;
 
 class JournalEntry extends React.Component {
   constructor(props) {
@@ -129,7 +120,7 @@ class JournalEntry extends React.Component {
 
     client
       .query({
-        query: GET_CURRENT_USERID
+        query: GET_CURRENT_USER_QUERY
       })
       .then(response => {
         this.setState({ currentUser: response.data.getCurrentUser.id });
