+++ client/src/graphql/queries.js
@@ -1,5 +1,6 @@
 import { gql } from "apollo-boost";
 
+// User
 export const GET_FOOD_ENTRIES_BY_USER_QUERY = gql`
   query getFoodEntriesByUserId($userId: ID!) {
     getFoodEntriesByUserId(userId: $userId) {
@@ -45,6 +46,23 @@ export const GET_CURRENT_USER_QUERY = gql`
   }
 `;
 
+export const GET_USER_BY_ID = gql`
+  query getUserById($userId: ID!){
+    getUserById(userId: $userId){
+      id
+      weight
+    }
+  }
+`;
+
+export const USER_EXIST_QUERY = gql`
+  query getUserBy($param: String!, $value: String!) {
+    getUserBy(param: $param, value: $value) {
+      email
+    }
+  }
+`;
+
 export const SEARCH_USER_BY_EMAIL = gql`
   query getUserBy($param: String!, $value: String!) {
     getUserBy(param: $param, value: $value) {
@@ -59,6 +77,7 @@ export const SEARCH_USER_BY_EMAIL = gql`
   }
 `;
 
+// Weight
 export const GET_WEIGHT_ENTRIES_QUERY = gql`
   query getWeightEntriesByUserId($userId: ID!) {
     getWeightEntriesByUserId(userId: $userId) {
@@ -69,6 +88,7 @@ export const GET_WEIGHT_ENTRIES_QUERY = gql`
   }
 `;
 
+// Exercise
 export const GET_EXERCISE_ENTRIES_QUERY = gql`
   query getExerciseEntriesByUserId($userId: ID!) {
     getExerciseEntriesByUserId(userId: $userId) {
@@ -76,44 +96,6 @@ export const GET_EXERCISE_ENTRIES_QUERY = gql`
       exerciseName
       caloriesBurned
       id
-    }
-  }
-`;
-
-export const GET_ALL_FOOD = gql`
-  query {
-    getFoods {
-      id
-      edamam_id
-    }
-  }
-`;
-
-export const GET_CURRENT_USERID = gql`
-  query {
-    getCurrentUser {
-      id
-      userType
-    }
-  }
-`;
-
-export const GET_USER_BY_ID = gql`
-  query getUserById($userId: ID!){
-    getUserById(userId: $userId){
-      id
-      weight
-    }
-  }
-`;
-
-
-
-
-export const USER_EXIST_QUERY = gql`
-  query getUserBy($param: String!, $value: String!) {
-    getUserBy(param: $param, value: $value) {
-      email
     }
   }
 `;
@@ -129,6 +111,49 @@ export const EXER_QUERY = gql`
   }
 `;
 
+//Food
+export const GET_ALL_FOOD = gql`
+  query {
+    getFoods {
+      id
+      edamam_id
+    }
+  }
+`;
+export const GET_FOOD_BY_ID = gql`
+  query getFoodById($foodId: ID!) {
+    getFoodById(foodId: $foodId) {
+      id
+      edamam_id
+    }
+  }
+`;
+
+//Food Entry Query
+export const FOOD_ENTRY_QUERY = gql`
+  query getFoodEntry($userId: ID!) {
+    getFoodEntriesByUserId(userId: $userId) {
+      id
+      date
+      servingQty
+      food_id {
+        id
+        foodName
+        caloriesPerServ
+        fats
+        proteins
+        carbs
+        edamam_id
+      }
+      meal_category_id {
+        id
+        mealCategoryName
+      }
+    }
+  }
+`;
+
+// Messages
 export const GET_MESSAGES_QUERY = gql`
   query getMessagesBy($param: String!, $value: String!) {
     getMessagesBy(param: $param, value: $value) {
@@ -151,6 +176,7 @@ export const GET_MESSAGES_QUERY = gql`
   }
 `;
 
+// Coaches
 export const GET_TRAINEES = gql`
   query getTrainees($coach_id: ID!){
     getTrainees(coach_id: $coach_id){
@@ -178,3 +204,22 @@ export const GET_COACHES = gql`
     }
   }
 `
+
+// Subscription
+export const GET_RECENT_BILLING = gql`
+  query getRecentBilling($id: ID!){
+    getRecentBilling(id: $id){
+      date
+    }
+  }
+`;
+
+export const GET_BILLING_HISTORY = gql`
+query getBillingHistory($id: ID!){
+    getBillingHistory(id: $id){
+        id
+        date
+        amount_paid
+    }
+  }
+`;
